watch for Game End after all
We can't always rely on the length specified by UBJSON for the `raw`
element, as it will be zero for in-progress replays. So we still need to
stop once we hit Game End, even though the spec doesn't currently
guarantee this will be the last event in the stream.

Partially reverts e6aa8b6d27c987874ab127626404ed5b1081b82f.